Move PropagateEquals to top level class
